Doctest that doesn't work in Python 2
Doctest seems to work by comparing the string representations of the
values, rather than the values themselves, so in Python 2, the three
lines of this doctest example that have string values all fail because
e.g. it says that it expects b'Hello, but got 'Hello', even though in
Python 2 `b'Hello,' == 'Hello,'` is True.